move template creator to Common\Templates\Abstractions
