+++ src/ArmTemplates/Common/Templates/Abstractions/TemplateCreator.cs
@@ -1,9 +1,8 @@
 ﻿using System.Collections.Generic;
 using Microsoft.Azure.Management.ApiManagement.ArmTemplates.Common.Constants;
-using Microsoft.Azure.Management.ApiManagement.ArmTemplates.Common.Templates.Abstractions;
 using Microsoft.Azure.Management.ApiManagement.ArmTemplates.Extractor.Models;
 
-namespace Microsoft.Azure.Management.ApiManagement.ArmTemplates.Creator.TemplateCreators
+namespace Microsoft.Azure.Management.ApiManagement.ArmTemplates.Common.Templates.Abstractions
 {
     public static class TemplateCreator
     {
